+++ core/src/syscall/precompiles/sha512/compress/mod.rs
@@ -92,7 +92,7 @@ pub const SHA512_COMPRESS_K: [u64; 80] = [
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Sha512CompressEvent {
-    pub lookup_id: usize,
+    pub lookup_id: u128,
     pub shard: u32,
     pub channel: u32,
     pub clk: u32,
